requirements.txt
pyzmq
pymongo>=3.0
mock
voluptuous
requests
responses
nose
six
aiozmq
aiohttp
